created custom hook to handle event click outside div
