+++ website/frontend/src/pages/CleanAir/CleanAirPublications.js
@@ -1,10 +1,9 @@
-import React, { useEffect, useState, useRef } from 'react';
+import React, { useEffect, useState } from 'react';
 import SEO from 'utilities/seo';
-import { useInitScrollTop } from 'utilities/customHooks';
+import { useInitScrollTop, useClickOutside } from 'utilities/customHooks';
 import { useDispatch, useSelector } from 'react-redux';
 import { setActiveResource } from 'reduxStore/CleanAirNetwork/CleanAir';
 import { ReportComponent } from 'components/CleanAir';
-import { getAllCleanAirApi } from 'apis/index';
 import { useTranslation } from 'react-i18next';
 import { RegisterSection, IntroSection, RotatingLoopIcon } from 'components/CleanAir';
 import ResourceImage from 'assets/img/cleanAir/resource.png';
@@ -13,14 +12,13 @@ import Slide from '@mui/material/Slide';
 import KeyboardDoubleArrowRightIcon from '@mui/icons-material/KeyboardDoubleArrowRight';
 import KeyboardDoubleArrowLeftIcon from '@mui/icons-material/KeyboardDoubleArrowLeft';
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
-import { fetchCleanAirData } from 'reduxStore/CleanAirNetwork/CleanAir';
 
 const ITEMS_PER_PAGE = 3;
 
 const CleanAirPublications = () => {
   useInitScrollTop();
   const { t } = useTranslation();
-  const filterRef = useRef(null);
+  const filterRef = useClickOutside(() => setFilter(false));
   const dispatch = useDispatch();
   const [openfilter, setFilter] = useState(false);
   const cleanAirResources = useSelector((state) => state.cleanAirData.airData);
@@ -52,19 +50,6 @@ const CleanAirPublications = () => {
   const researchPublicationData = cleanAirResources.filter(
     (resource) => resource.resource_category === 'research_publication'
   );
-
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (filterRef.current && !filterRef.current.contains(event.target)) {
-        setFilter(false);
-      }
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [filterRef]);
 
   const handleFilterSelect = (filter) => {
     dispatch(setActiveResource(filter));
